// prisma/generated/type-graphql/resolvers/crud/Invitations/args/UpsertOneInvitationsArgs.d.ts
import { InvitationsCreateInput } from "../../../inputs/InvitationsCreateInput";
import { InvitationsUpdateInput } from "../../../inputs/InvitationsUpdateInput";
import { InvitationsWhereUniqueInput } from "../../../inputs/InvitationsWhereUniqueInput";
export declare class UpsertOneInvitationsArgs {
    where: InvitationsWhereUniqueInput;
    create: InvitationsCreateInput;
    update: InvitationsUpdateInput;
}