// prisma/generated/type-graphql/models/Invitations.ts
import * as TypeGraphQL from "type-graphql";
import * as GraphQLScalars from "graphql-scalars";
import { Prisma } from "@prisma/client";
import { DecimalJSScalar } from "../scalars";

@TypeGraphQL.ObjectType("Invitations", {
  isAbstract: true
})
export class Invitations {
  @TypeGraphQL.Field(_type => GraphQLScalars.BigIntResolver, {
    nullable: false
  })
  id!: bigint;

  @TypeGraphQL.Field(_type => Date, {
    nullable: true
  })
  created_at?: Date | null;

  @TypeGraphQL.Field(_type => Date, {
    nullable: true
  })
  invoiceDate?: Date | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  email?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  billingAddress?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  shippingAddress?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  orderNumber?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  grandTotalPurchased?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  paymentMethod?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  status?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  media?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  content?: string | null;
}